Server bug fixes and global fixes
Server bug fixes with host now sending all players back to main menu if disconnecting. This should introduce a fix to players being incorrectly spawned in when not meant to when a player disconnects also. This fixes a bug where Global.players stored data of players that did not exist