Flinch, UI, and NoScope improvements
Flinch has been added for when a player is hit
UI improved and menu is scalable with resolution.
Also added quit on server joiner and leave match on choose class.
No scope improvement where when noscoping the bullet now comes from a raycast on the gun to give better clarity on tracer shots